refactor: change register of the files
